+++ src/idc/writer/objdet/__init__.py
@@ -1,5 +1,6 @@
 from ._adams import AdamsObjectDetectionWriter
 from ._coco import COCOObjectDetectionWriter
+from ._instancepng import InstancePngObjectDetectionWriter
 from ._opex import OPEXObjectDetectionWriter
 from ._roi import ROIObjectDetectionWriter
 from ._voc import VOCObjectDetectionWriter