+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   palette index for each object instance
 - added `PNG_EXTENSIONS` constant
 - added `imgseg_from_instancepng`, `objdet_from_instancepng`, `contours_to_objdet` methods
+- added `to-instance-png-od` writer that outputs the object annotations of a particular label
+  as separate palette indices
 
 
 0.1.0 (2025-10-31)